removed custom CORS endpoints & migrated to EnableWebMvc-managed web config
