+++ src/main/kotlin/com/idan/pokemon_hub/controller/PokemonController.kt
@@ -10,24 +10,6 @@ import org.springframework.web.bind.annotation.*
 @RequestMapping("/pokemons")
 class PokemonController(private val pokemonService: PokemonService) {
 
-    @RequestMapping(value = ["/"], method = [RequestMethod.OPTIONS])
-    fun optionsAll(): ResponseEntity<String> {
-        val msg = "The available method for all Pokémons is: GET (retrieve)."
-        return ResponseEntity
-            .ok()
-            .allow(HttpMethod.GET)
-            .body(msg)
-    }
-
-    @RequestMapping(value = ["/{pokedex}"], method = [RequestMethod.OPTIONS])
-    fun optionsSingular(@PathVariable pokedex: Long): ResponseEntity<String> {
-        val msg = "The available methods for a specific Pokémon are: GET (retrieve), PUT (update), DELETE (delete)."
-        return ResponseEntity
-            .ok()
-            .allow(HttpMethod.GET, HttpMethod.PUT, HttpMethod.DELETE)
-            .body(msg)
-    }
-
     @GetMapping
     fun getAll(): List<Pokemon> {
         return pokemonService.getAll()
